web/tasty-fare: changes refresh strategy
Will now refresh correctly in the face of devices going to sleep.
Previously it used setTimeout which doesn't correctly handle sleep.
While setInterval doesn't either, by checking on a shorter interval
to see if the timeout has elapsed we're more likely to notice that
the device slept and woke up, and refresh data.